Pin dependency versions
shinywidgets 0.3.4 is required.
Newer versions lead to ""AttributeError: 'NoneType' object has no attribute 'comm_id'"" errors.
This is caused because the ipywidget comm is closed by the reactive system of shiny.
We would need to replace the accesses to map.add and map.remove by a non-imperative pattern
in order to be able to upgrade shinywidgets.